create channelManagement folder and move epic and banner tests into it
