+++ public/src/components/channelManagement/epicTests/articlesViewedEditor.tsx
@@ -9,8 +9,8 @@ import {
   RadioGroup,
   Radio,
 } from "@material-ui/core";
-import EditableTextField from "../helpers/editableTextField"
-import {onFieldValidationChange, ValidationStatus} from "../helpers/validation";
+import EditableTextField from "../../helpers/editableTextField"
+import {onFieldValidationChange, ValidationStatus} from "../../helpers/validation";
 import {ArticlesViewedSettings} from "./epicTestsForm";
 
 const isNumber = (value: string): boolean => !Number.isNaN(Number(value));
