+++ public/src/components/channelManagement/epicTests/epicTestEditor.tsx
@@ -1,35 +1,37 @@
-import React from 'react';
-import {Region} from '../../utils/models';
-import {ArticlesViewedSettings, UserCohort} from "../epicTests/epicTestsForm"; //toDo make all shared
+import React, { ReactNode } from 'react';
+import { Region } from '../../../utils/models';
+import {EpicTest, EpicVariant, UserCohort, MaxEpicViews, ArticlesViewedSettings} from "./epicTestsForm";
 import {
-  createStyles,
+  Checkbox,
   FormControl,
   FormControlLabel,
+  FormGroup,
+  Input,
   InputLabel,
+  ListItemText,
+  MenuItem,
   Radio,
   RadioGroup,
+  Select,
   Switch,
   Theme,
   Typography,
   WithStyles,
-  withStyles,
-  FormGroup,
-  Checkbox,
+  createStyles,
+  withStyles
 } from "@material-ui/core";
-import {onFieldValidationChange, isNumber} from '../helpers/validation';
+import EditableTextField from '../helpers/editableTextField';
+import EpicTestVariantsList from './epicTestVariantsList';
+import MaxEpicViewsEditor from './maxEpicViewsEditor';
+import {onFieldValidationChange, ValidationStatus} from '../helpers/validation';
 import ButtonWithConfirmationPopup from '../helpers/buttonWithConfirmationPopup';
 import DeleteSweepIcon from '@material-ui/icons/DeleteSweep';
 import ArchiveIcon from '@material-ui/icons/Archive';
-import {TestEditorState} from "../epicTests/epicTestEditor";
-import {BannerTest, BannerVariant} from "./bannerTestsForm";
-import TargetRegionsSelector from "../epicTests/targetRegionsSelector";
-import ArticlesViewedEditor, {defaultArticlesViewedSettings} from "../epicTests/articlesViewedEditor";
-import {articleCountTemplate} from "../epicTests/epicTestVariantEditor";
-import NewNameCreator from "../epicTests/newNameCreator";
-import BannerTestVariantsList from "./bannerTestVariantsList";
-import UserCohortSelector from ".././epicTests/userCohortSelector";
-import EditableTextField from "../helpers/editableTextField"
-
+import {articleCountTemplate, countryNameTemplate} from './epicTestVariantEditor';
+import ArticlesViewedEditor, {defaultArticlesViewedSettings} from './articlesViewedEditor';
+import NewNameCreator from './newNameCreator';
+import EpicTypeComponent, {EpicType} from './epicTypeComponent';
+import TargetRegionsSelector from './targetRegionsSelector';
 
 const styles = ({ spacing, typography}: Theme) => createStyles({
   container: {
@@ -98,20 +100,17 @@ const styles = ({ spacing, typography}: Theme) => createStyles({
     marginRight: spacing(6),
     fontSize: typography.pxToRem(18),
   },
-  hr: {
-    width: '100%',
-  }
 });
 
-const copyHasTemplate = (test: BannerTest, template: string): boolean => test.variants.some(variant =>
+const copyHasTemplate = (test: EpicTest, template: string): boolean => test.variants.some(variant =>
   variant.heading && variant.heading.includes(template) ||
-  variant.body.includes(template)
+  variant.paragraphs.some(para => para.includes(template))
 );
 
-interface BannerTestEditorProps extends WithStyles<typeof styles> {
-  test?: BannerTest,
+interface EpicTestEditorProps extends WithStyles<typeof styles> {
+  test?: EpicTest,
   hasChanged: boolean,
-  onChange: (updatedTest: BannerTest) => void,
+  onChange: (updatedTest: EpicTest) => void,
   onValidationChange: (isValid: boolean) => void,
   visible: boolean,
   editMode: boolean,
@@ -122,12 +121,16 @@ interface BannerTestEditorProps extends WithStyles<typeof styles> {
   isNew: boolean,
   testNames: string[],
   testNicknames: string[],
-  createTest: (newTest: BannerTest) => void
+  createTest: (newTest: EpicTest) => void
+}
+//todo move to shared
+export interface TestEditorState {
+  validationStatus: ValidationStatus
 }
 
 const areYouSure = `Are you sure? This can't be undone without cancelling entire edit session!`;
 
-class BannerTestEditor extends React.Component<BannerTestEditorProps, TestEditorState> {
+class EpicTestEditor extends React.Component<EpicTestEditorProps, TestEditorState> {
 
   state: TestEditorState = {
     validationStatus: {}
@@ -137,7 +140,7 @@ class BannerTestEditor extends React.Component<BannerTestEditorProps, TestEditor
     return this.props.editMode && !this.props.isDeleted && !this.props.isArchived;
   }
 
-  getArticlesViewedSettings = (test: BannerTest): ArticlesViewedSettings | undefined => {
+  getArticlesViewedSettings = (test: EpicTest): ArticlesViewedSettings | undefined => {
     if (!!test.articlesViewedSettings) {
       return test.articlesViewedSettings;
     }
@@ -147,13 +150,14 @@ class BannerTestEditor extends React.Component<BannerTestEditorProps, TestEditor
     return undefined;
   }
 
-  updateTest = (update: (test: BannerTest) => BannerTest) => {
+  updateTest = (update: (test: EpicTest) => EpicTest) => {
     if (this.props.test) {
       const updatedTest = update(this.props.test);
 
       this.props.onChange({
         ...updatedTest,
         // To save dotcom from having to work this out
+        hasCountryName: copyHasTemplate(updatedTest, countryNameTemplate),
         articlesViewedSettings: this.getArticlesViewedSettings(updatedTest),
       })
     }
@@ -161,7 +165,7 @@ class BannerTestEditor extends React.Component<BannerTestEditorProps, TestEditor
 
   copyTest = (newTestName: string, newTestNickname: string): void => {
     if (this.props.test) {
-      const newTest: BannerTest = {
+      const newTest: EpicTest = {
         ...this.props.test,
         name: newTestName,
         nickname: newTestNickname,
@@ -170,7 +174,7 @@ class BannerTestEditor extends React.Component<BannerTestEditorProps, TestEditor
     }
   }
 
-  onVariantsChange = (updatedVariantList: BannerVariant[]): void => {
+  onVariantsChange = (updatedVariantList: EpicVariant[]): void => {
     if (this.props.test) {
       this.updateTest(test => ({...test, "variants": updatedVariantList}));
     }
@@ -186,15 +190,21 @@ class BannerTestEditor extends React.Component<BannerTestEditorProps, TestEditor
     this.updateTest(test => ({...test, [fieldName]: updatedBool}));
   }
 
-  onUserCohortChange = (selectedCohort: UserCohort): void => {
+  onUserCohortChange = (event: React.ChangeEvent<{}>, value: string): void => {
+    let selectedCohort = value as UserCohort;
     this.updateTest(test => ({...test, "userCohort": selectedCohort}));
+  }
+
+  onEpicTypeChange = (epicType: EpicType): void => {
+    const isLiveBlog = epicType === 'LiveBlog';
+    this.updateTest(test => ({...test, "isLiveBlog": isLiveBlog}))
   }
 
   onTargetRegionsChange = (selectedRegions: Region[]): void => {
     this.updateTest(test => ({...test, 'locations': selectedRegions}));
   }
 
-  renderBottomButtons = (test: BannerTest) => (
+  renderBottomButtons = (test: EpicTest) => (
     <div className={this.props.classes.buttons}>
       <div className={this.props.classes.button}>
         <ButtonWithConfirmationPopup
@@ -226,7 +236,7 @@ class BannerTestEditor extends React.Component<BannerTestEditorProps, TestEditor
     </div>
   )
 
-  renderEditor = (test: BannerTest): React.ReactNode => {
+  renderEditor = (test: EpicTest): React.ReactNode => {
     const {classes} = this.props;
 
     const statusText = () => {
@@ -256,39 +266,58 @@ class BannerTestEditor extends React.Component<BannerTestEditorProps, TestEditor
 
         <hr />
 
+        <EpicTypeComponent
+          epicType={test.isLiveBlog ? 'LiveBlog' : 'Standard'}
+          isEditable={this.isEditable()}
+          onEpicTypeChange={this.onEpicTypeChange}
+        />
+
         <Typography variant={'h4'} className={classes.boldHeading}>Variants</Typography>
         <div>
-            <BannerTestVariantsList
-              variants={test.variants}
-              onVariantsListChange={this.onVariantsChange}
-              testName={test.name}
-              editMode={this.isEditable()}
-              onValidationChange={onFieldValidationChange(this)('variantsList')}
-            />
+          <EpicTestVariantsList
+            variants={test.variants}
+            onVariantsListChange={this.onVariantsChange}
+            testName={test.name}
+            editMode={this.isEditable()}
+            onValidationChange={onFieldValidationChange(this)('variantsList')}
+          />
         </div>
 
+        <Typography variant={'h4'} className={classes.boldHeading}>Target content</Typography>
+
         <div>
-          <Typography variant={'h4'} className={classes.boldHeading}>Display rules</Typography>
           <EditableTextField
-            text={test.minArticlesBeforeShowingBanner.toString()}
-            onSubmit={(pageViews: string) =>
-              this.updateTest(test => ({ ...test, minArticlesBeforeShowingBanner: Number(pageViews)}))
-            }
-            label={'Show the banner on'}
-            helperText="Must be a number"
-            editEnabled={this.props.editMode}
-            validation={
-              {
-                getError: (value: string) => isNumber(value) ? null : 'Must be a number',
-                onChange: onFieldValidationChange(this)('minArticlesBeforeShowingBanner')
-              }
-            }
-            isNumberField
-        />
-        <Typography>page views</Typography>
-        </div>
+            text={test.tagIds.join(",")}
+            onSubmit={this.onListChange('tagIds')}
+            label="Target tags"
+            helperText="Format: environment/wildlife,business/economics"
+            editEnabled={this.isEditable()}
+          />
 
-        <div>
+          <EditableTextField
+            text={test.sections.join(",")}
+            onSubmit={this.onListChange('sections')}
+            label="Target sections"
+            helperText="Format: environment,business"
+            editEnabled={this.isEditable()}
+          />
+
+          <EditableTextField
+            text={test.excludedTagIds.join(",")}
+            onSubmit={this.onListChange('excludedTagIds')}
+            label="Excluded tags"
+            helperText="Format: environment/wildlife,business/economics"
+            editEnabled={this.isEditable()}
+          />
+
+          <EditableTextField
+            text={test.excludedSections.join(",")}
+            onSubmit={this.onListChange('excludedSections')}
+            label="Excluded sections"
+            helperText="Format: environment,business"
+            editEnabled={this.isEditable()}
+          />
+
           <Typography variant={'h4'} className={classes.boldHeading}>Target audience</Typography>
 
           <TargetRegionsSelector
@@ -297,13 +326,54 @@ class BannerTestEditor extends React.Component<BannerTestEditorProps, TestEditor
             isEditable={this.isEditable()}
           />
 
-          <UserCohortSelector
-            cohort={test.userCohort}
-            onCohortsUpdate={this.onUserCohortChange}
-            isEditable={this.isEditable()}
+          <FormControl
+            className={classes.formControl}>
+              <InputLabel
+                className={classes.selectLabel}
+                shrink
+                htmlFor="user-cohort">
+                  Supporter status
+              </InputLabel>
+              <RadioGroup
+                className={classes.radio}
+                value={test.userCohort}
+                onChange={this.onUserCohortChange}
+              >
+                {Object.values(UserCohort).map(cohort =>
+                  <FormControlLabel
+                    value={cohort}
+                    key={cohort}
+                    control={<Radio />}
+                    label={cohort}
+                    disabled={!this.isEditable()}
+                  />
+                )}
+              </RadioGroup>
+          </FormControl>
+
+          <Typography variant={'h4'} className={this.props.classes.boldHeading}>View frequency settings</Typography>
+
+          <FormControlLabel
+            control={
+              <Switch
+                checked={test.useLocalViewLog}
+                onChange={this.onSwitchChange('useLocalViewLog')}
+                disabled={!this.isEditable()}
+              />
+            }
+            label={`Use private view counter for this test (instead of the global one)`}
           />
 
-          <hr className={classes.hr} />
+          <MaxEpicViewsEditor
+            test={test}
+            editMode={this.isEditable()}
+            onChange={(alwaysAsk: boolean, maxEpicViews: MaxEpicViews) =>
+              this.updateTest(test => ({ ...test, alwaysAsk, maxViews: maxEpicViews }))
+            }
+            onValidationChange={onFieldValidationChange(this)('maxViews')}
+          />
+
+          <Typography variant={'h4'} className={this.props.classes.boldHeading}>Article count</Typography>
           <ArticlesViewedEditor
             articlesViewedSettings={test.articlesViewedSettings}
             editMode={this.isEditable()}
@@ -327,4 +397,4 @@ class BannerTestEditor extends React.Component<BannerTestEditorProps, TestEditor
   }
 }
 
-export default withStyles(styles)(BannerTestEditor);
+export default withStyles(styles)(EpicTestEditor);
