+++ public/src/main.tsx
@@ -5,7 +5,7 @@ import {BrowserRouter as Router, Route} from 'react-router-dom';
 import Switchboard from './components/switchboard';
 import ContributionTypesForm from './components/contributionTypes';
 import AmountsForm from './components/amounts/amounts';
-import EpicTestsForm from './components/epicTests/epicTestsForm';
+import EpicTestsForm from './components/channelManagement/epicTests/epicTestsForm';
 import {createStyles, Theme, WithStyles, withStyles} from '@material-ui/core/styles';
 import CssBaseline from '@material-ui/core/CssBaseline';
 import AppBar from '@material-ui/core/AppBar';
@@ -14,7 +14,7 @@ import NavDrawer from "./components/drawer";
 import Toolbar from "@material-ui/core/Toolbar";
 import Typography from "@material-ui/core/Typography";
 import IndexPage from "./components/indexPage";
-import BannerTestsForm from "./components/bannerTests/bannerTestsForm";
+import BannerTestsForm from "./components/channelManagement/bannerTests/bannerTestsForm";
 
 const styles = ({ palette, spacing, mixins, typography, transitions }: Theme) => createStyles({
   root: {
